+++ client/scripts/components/detail/detail-researcher.jsx
@@ -60,7 +60,7 @@ var Template = React.createClass({
                   onAdd={this.props.onRelationshipAdd.bind(null, 'functions')}
                   onRemove={this.props.onRelationshipRemove.bind(null, 'functions')}
                   field={'functions'}
-                  routeName='phenomenon'
+                  routeName='FM'
                   title='Outcomes'
                   fieldName='Outcome'
                   titleField={'name'} />
