+++ client/scripts/components/detail/productDetail.jsx
@@ -124,10 +124,10 @@ var ProductDetail = React.createClass({
               <RelationshipList items={detail.conditions} editable={this.state.editable} onAdd={this.onRelationshipAdd.bind(null, 'conditions')} onRemove={this.onRelationshipRemove.bind(null, 'conditions')} field={'conditions'} routeName='condition' title='Context' fieldName='Context' titleField={'name'} />
             </Col>
             <Col xs={6} sm={4}>
-              <RelationshipList items={detail.mechanisms} editable={this.state.editable} onAdd={this.onRelationshipAdd.bind(null, 'mechanisms')} onRemove={this.onRelationshipRemove.bind(null, 'mechanisms')} field={'mechanisms'} routeName='phenomenon' title='Mechanisms' fieldName='Mechanism' titleField={'name'} />
+              <RelationshipList items={detail.mechanisms} editable={this.state.editable} onAdd={this.onRelationshipAdd.bind(null, 'mechanisms')} onRemove={this.onRelationshipRemove.bind(null, 'mechanisms')} field={'mechanisms'} routeName='FM' title='Mechanisms' fieldName='Mechanism' titleField={'name'} />
             </Col>
             <Col xs={6} sm={4}>
-              <RelationshipList items={detail.outcomes} editable={this.state.editable} onAdd={this.onRelationshipAdd.bind(null, 'outcomes')} onRemove={this.onRelationshipRemove.bind(null, 'outcomes')} field={'outcomes'} routeName='phenomenon' title='Outcomes' fieldName='Outcome' titleField={'name'} />
+              <RelationshipList items={detail.outcomes} editable={this.state.editable} onAdd={this.onRelationshipAdd.bind(null, 'outcomes')} onRemove={this.onRelationshipRemove.bind(null, 'outcomes')} field={'outcomes'} routeName='FM' title='Outcomes' fieldName='Outcome' titleField={'name'} />
             </Col>
             <Col xs={12} sm={8}>
               <RelationshipList items={detail.strategies} editable={this.state.editable} onAdd={this.onRelationshipAdd.bind(null, 'strategies')} onRemove={this.onRelationshipRemove.bind(null, 'strategies')} field={'strategies'} routeName='strategy' title='Inspired By' fieldName='Biological Strategy' titleField={'name'} />
