+++ client/scripts/components/detail/strategyDetail.jsx
@@ -105,10 +105,10 @@ var StrategyDetail = React.createClass({
                            <RelationshipList items={this.state.object.conditions} editable={this.state.editable} titleField='name' onAdd={this.onRelationshipAdd.bind(null, 'conditions')} onRemove={this.onRelationshipRemove.bind(null, 'conditions')} field={'conditions'} routeName={'condition'} title={'Context'} fieldName={'Context'}/>
             		    </Col>
             		    <Col xs={6} sm={4}>
-                            <RelationshipList items={this.state.object.mechanisms} editable={this.state.editable} titleField='name' onAdd={this.onRelationshipAdd.bind(null,'mechanisms')} onRemove={this.onRelationshipRemove.bind(null, 'mechanisms')} field={'functions'} routeName={'phenomenon'} title={'Mechanisms'} fieldName={'Mechanisms'}/>
+                            <RelationshipList items={this.state.object.mechanisms} editable={this.state.editable} titleField='name' onAdd={this.onRelationshipAdd.bind(null,'mechanisms')} onRemove={this.onRelationshipRemove.bind(null, 'mechanisms')} field={'functions'} routeName={'FM'} title={'Mechanisms'} fieldName={'Mechanisms'}/>
             		    </Col>
             		    <Col xs={6} sm={4}>
-                          <RelationshipList items={this.state.object.functions} editable={this.state.editable} titleField='name' onAdd={this.onRelationshipAdd.bind(null, 'functions')} onRemove={this.onRelationshipRemove.bind(null, 'functions')} field={'functions'} routeName={'phenomenon'} title={'Outcomes'} fieldName={'Outcomes'}/>
+                          <RelationshipList items={this.state.object.functions} editable={this.state.editable} titleField='name' onAdd={this.onRelationshipAdd.bind(null, 'functions')} onRemove={this.onRelationshipRemove.bind(null, 'functions')} field={'functions'} routeName={'FM'} title={'Outcomes'} fieldName={'Outcomes'}/>
             		    </Col>
           		</Row>
          	    </Grid>
