+++ server/models/context.js
@@ -1,4 +1,4 @@
-// Phenomenon model
+// FM model
 'use strict';
 
 var Model = require('./model.js');
